Pipeline works again as v2
